Refactor, ready for testing

+++ shared_headers/cpu_shared_info.h
@@ -1,13 +1,13 @@
 #ifndef CPU_SHARED_INFO_H
 #define CPU_SHARED_INFO_H
 
-#include <stdint.h>
 #include <winioctl.h>
+#include <wdm.h>
 
 typedef struct _CPU_DATA {
-    int16_t temp;  // There exists some beastly AMD cpu with 192 cores
-    uint64_t cpu_load; 
-    uint32_t cpu_id;
+    WORD temp;  // There exists some beastly AMD cpu with 192 cores
+    ULONGLONG cpu_load; 
+    ULONG cpu_id;
 } CPU_DATA, *PCPU_DATA;
 
 typedef struct _CPU_DATA_HEADER {
